Refactor shim so that it can be called without url parameters
Addresses #6
